requirements_venv.txt
aiobotocore==2.4.0
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
asyncssh==2.12.0
attrs==22.1.0
bleach==5.0.1
boto3==1.24.59
botocore==1.27.59
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.2.0
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
docutils==0.19
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.1
hydra-core==1.2.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
jaraco.classes==3.2.3
jmespath==1.0.1
joblib==1.2.0
keyring==23.9.3
kiwisolver==1.4.4
kneed==0.8.1
kubernetes-asyncio==24.2.2
matplot==0.1.9
matplotlib==3.6.1
meadowrun==0.2.11
more-itertools==9.0.0
multidict==6.0.2
numpy==1.23.4
omegaconf==2.2.3
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pkginfo==1.8.3
protobuf==3.20.3
psutil==5.9.3
pycparser==2.21
Pygments==2.13.0
pyloco==0.0.139
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
readme-renderer==37.2
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
SimpleWebSocketServer==0.1.2
six==1.16.0
threadpoolctl==3.1.0
twine==4.0.1
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.12
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.4.1
wrapt==1.14.1
yarl==1.8.1
zipp==3.10.0
